requirements.txt
beautifulsoup4
mysql-connector-python
pytest
python-dotenv
